+++ boards/arduino_nano_connect/examples/nano_blinky.rs
@@ -11,8 +11,6 @@
 #![no_main]
 
 use embedded_hal::digital::v2::OutputPin;
-// // Time handling traits
-use embedded_time::rate::*;
 
 // Ensure we halt the program on panic (if we don't mention this crate it won't
 // be linked)
@@ -64,7 +62,7 @@ fn main() -> ! {
 
     // The delay object lets us wait for specified amounts of time (in
     // milliseconds)
-    let mut delay = cortex_m::delay::Delay::new(core.SYST, clocks.system_clock.freq().integer());
+    let mut delay = cortex_m::delay::Delay::new(core.SYST, clocks.system_clock.freq().to_Hz());
 
     // The single-cycle I/O block controls our GPIO pins
     let sio = hal::Sio::new(pac.SIO);