+++ boards/rp-pico/examples/pico_pio_pwm.rs
@@ -17,9 +17,6 @@ use defmt_rtt as _;
 // The macro for our start-up function
 use rp_pico::entry;
 
-// Time handling traits
-use embedded_time::rate::*;
-
 // Ensure we halt the program on panic (if we don't mention this crate it won't
 // be linked)
 use panic_halt as _;
@@ -125,7 +122,7 @@ fn main() -> ! {
 
     // The delay object lets us wait for specified amounts of time (in
     // milliseconds)
-    let mut delay = cortex_m::delay::Delay::new(core.SYST, clocks.system_clock.freq().integer());
+    let mut delay = cortex_m::delay::Delay::new(core.SYST, clocks.system_clock.freq().to_Hz());
 
     let (mut pio0, sm0, _, _, _) = pac.PIO0.split(&mut pac.RESETS);