+++ boards/seeeduino-xiao-rp2040/examples/seeeduino_xiao_rp2040_blinky.rs
@@ -17,9 +17,6 @@ use seeeduino_xiao_rp2040::entry;
 use embedded_hal::digital::v2::OutputPin;
 use embedded_hal::PwmPin;
 
-// Time handling traits
-use embedded_time::rate::*;
-
 // Ensure we halt the program on panic (if we don't mention this crate it won't
 // be linked)
 use panic_halt as _;
@@ -85,7 +82,7 @@ fn main() -> ! {
 
     // The delay object lets us wait for specified amounts of time (in
     // milliseconds)
-    let mut delay = cortex_m::delay::Delay::new(core.SYST, clocks.system_clock.freq().integer());
+    let mut delay = cortex_m::delay::Delay::new(core.SYST, clocks.system_clock.freq().to_Hz());
 
     // Init PWMs
     let mut pwm_slices = hal::pwm::Slices::new(pac.PWM, &mut pac.RESETS);