+++ rp2040-hal/examples/adc.rs
@@ -20,7 +20,6 @@ use rp2040_hal as hal;
 // Some traits we need
 use core::fmt::Write;
 use embedded_hal::adc::OneShot;
-use embedded_time::fixed_point::FixedPoint;
 use rp2040_hal::Clock;
 
 // A shorter alias for the Peripheral Access Crate, which provides low-level
@@ -68,7 +67,7 @@ fn main() -> ! {
 
     // The delay object lets us wait for specified amounts of time (in
     // milliseconds)
-    let mut delay = cortex_m::delay::Delay::new(core.SYST, clocks.system_clock.freq().integer());
+    let mut delay = cortex_m::delay::Delay::new(core.SYST, clocks.system_clock.freq().to_Hz());
 
     // The single-cycle I/O block controls our GPIO pins
     let sio = hal::Sio::new(pac.SIO);