+++ rp2040-hal/examples/lcd_display.rs
@@ -23,7 +23,6 @@ use rp2040_hal as hal;
 use hd44780_driver as hd44780;
 
 // Some traits we need
-use embedded_time::fixed_point::FixedPoint;
 use rp2040_hal::clocks::Clock;
 
 // A shorter alias for the Peripheral Access Crate, which provides low-level
@@ -71,7 +70,7 @@ fn main() -> ! {
 
     // The delay object lets us wait for specified amounts of time (in
     // milliseconds)
-    let mut delay = cortex_m::delay::Delay::new(core.SYST, clocks.system_clock.freq().integer());
+    let mut delay = cortex_m::delay::Delay::new(core.SYST, clocks.system_clock.freq().to_Hz());
 
     // The single-cycle I/O block controls our GPIO pins
     let sio = hal::Sio::new(pac.SIO);